% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecospat.boyce.R
\name{ecospat.boyce}
\alias{ecospat.boyce}
\title{ecospat.boyce}
\usage{
ecospat.boyce(
  fit,
  obs,
  nclass = 0,
  window.w = "default",
  res = 100,
  PEplot = TRUE,
  rm.duplicate = TRUE,
  method = "spearman"
)
}
\arguments{
\item{fit}{A vector or a SpatRaster containing the predicted suitability values.}

\item{obs}{A vector containing the predicted suitability values of validation points,
or xy-coordinates (if fit is a RasterLayer/SpatRaster).}

\item{nclass}{Number of classes or a vector with class thresholds.
If nclass = 0, a moving window approach is used.}

\item{window.w}{Width of the moving window (default = 1/10 of suitability range).}

\item{res}{Resolution of the moving window (default = 100 focal points).}

\item{PEplot}{Logical. If TRUE, plot predicted/expected ratios.}

\item{rm.duplicate}{Logical. If TRUE, remove successive duplicated values.}

\item{method}{Correlation method used for the Boyce index ('spearman' or 'pearson').}
}
\value{
A list with F.ratio (Predicted/Expected), cor (Boyce index), and HS (suitability classes).
}
\description{
Updated ecospat.boyce (ecospat package) function for calculating Boyce index as in Hirzel et al. 2006
}
