% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.fit_class.R
\docType{class}
\name{nsdm.fit-class}
\alias{nsdm.fit-class}
\alias{nsdm.fit}
\title{Generate an S4 class to store fitted models}
\description{
Not to be called directly by the user
}
\author{
Philipp Brun (philipp.brun@wsl.ch)
}
